Add a nix-darwin module
This is somewhat but not entirely duplicative with the home manager
support. The difference is primarily for MacOS servers, for which there
is no LaunchAgent support (as launch agents are tied to user sessions).

This PR adds a nix-darwin module (configured similarly to the home
manager module) which is suitable for e.g. other launchd daemons.